feat: :sparkles: Chart in home, working using data from firebase
